+++ SSD-Major-Web-Project/ViewModels/ShoppingCartVM.cs
@@ -11,14 +11,14 @@ namespace SSD_Major_Web_Project.ViewModels
         public List<Product> Products { get; set; }
         public string? CouponCode { get; set; }
         [DisplayFormat(DataFormatString = "{0:C}")]
-        public double Subtotal { get; set; }
+        public decimal Subtotal { get; set; }
         [DisplayFormat(DataFormatString = "{0:C}")]
-        public double ShippingFee { get; set; }
+        public decimal ShippingFee { get; set; }
         [DisplayFormat(DataFormatString = "{0:C}")]
-        public double Taxes { get; set; }
+        public decimal Taxes { get; set; }
         // [DisplayFormat(DataFormatString = "{0:C}")]
 
-        public double GrandTotal { get; set; }
+        public decimal GrandTotal { get; set; }
         public string Currency { get; set; }
         public string CurrencySymbol { get; set; }
         //public byte[]? ImageByteArray { get; set; }
